Refactor and harden genome consensus output
